Relocate required_access tests
